requirements.txt
openai>=1.0.0
mcrcon>=0.7.0
gradio>=4.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
